+++ packages/audit-management-widget/src/lib/widget/state/initialState.ts
@@ -1,0 +1,11 @@
+import { State } from './types';
+
+export const initialState: State = {
+  auditList: {
+    data: [],
+    loading: false,
+    error: null,
+  },
+  searchParams: { text: '', sort: [] },
+  selectedAuditId: '',
+};