+++ packages/web-component/src/lib/types.ts
@@ -31,6 +31,7 @@ export interface ScreenState {
   inputs?: Record<string, string>; // Backward compatibility
   lastAuth?: LastAuthState;
   totp?: { image?: string; provisionUrl?: string };
+  notp?: { image?: string; redirectUrl?: string };
 }
 
 export type FlowState = {
@@ -42,6 +43,7 @@ export type FlowState = {
   executionId: string;
   action: string;
   redirectTo: string;
+  openInNewTabUrl?: string;
   redirectUrl: string;
   screenId: string;
   screenState: ScreenState;