+++ packages/web-component/test/descope-wc.test.ts
@@ -96,6 +96,7 @@ Object.defineProperty(window, 'location', {
   value: new URL(window.location.origin),
 });
 window.location.assign = jest.fn();
+window.open = jest.fn();
 
 Object.defineProperty(window, 'PublicKeyCredential', { value: TestClass });
 
@@ -972,12 +973,19 @@ describe('web-component', () => {
     });
   });
 
-  it('should update totp link href according to screen state', async () => {
+  it('should update totp and notp link href according to screen state', async () => {
     startMock.mockReturnValue(
-      generateSdkResponse({ screenState: { totp: { provisionUrl: 'url1' } } }),
+      generateSdkResponse({
+        screenState: {
+          totp: { provisionUrl: 'url1' },
+          notp: { redirectUrl: 'url2' },
+        },
+      }),
     );
 
-    pageContent = `<div>Loaded1</div><descope-link data-type="totp-link">Provision URL</descope-link>`;
+    pageContent = `<div>Loaded1</div>
+      <descope-link data-type="totp-link">Provision URL</descope-link>
+      <descope-link data-type="notp-link">Redirect URL</descope-link>`;
 
     document.body.innerHTML = `<h1>Custom element test</h1> <descope-wc flow-id="otpSignInEmail" project-id="1"></descope-wc>`;
 
@@ -988,6 +996,9 @@ describe('web-component', () => {
 
     const totpLink = screen.getByShadowText('Provision URL');
     expect(totpLink).toHaveAttribute('href', 'url1');
+
+    const notpLink = screen.getByShadowText('Redirect URL');
+    expect(notpLink).toHaveAttribute('href', 'url2');
   });
 
   it('should disable webauthn buttons when its not supported in the browser', async () => {
@@ -1369,6 +1380,35 @@ describe('web-component', () => {
         timeout: WAIT_TIMEOUT,
       },
     );
+  });
+
+  it('When response has "openInNewTabUrl" it opens the URL in a new window', async () => {
+    nextMock.mockReturnValueOnce(
+      generateSdkResponse({
+        openInNewTabUrl: 'https://loremipsumurl.com',
+      }),
+    );
+
+    pageContent = '<span>It works!</span>';
+    window.location.search = `?${URL_RUN_IDS_PARAM_NAME}=0_0&${URL_CODE_PARAM_NAME}=code1`;
+    document.body.innerHTML = `<h1>Custom element test</h1> <descope-wc flow-id="versioned-flow" project-id="1"></descope-wc>`;
+
+    // Make sure url is opened in a new tab
+    await waitFor(
+      () =>
+        expect(window.open).toHaveBeenCalledWith(
+          'https://loremipsumurl.com',
+          '_blank',
+        ),
+      {
+        timeout: WAIT_TIMEOUT,
+      },
+    );
+
+    // Should also show the screen
+    await waitFor(() => screen.findByShadowText('It works!'), {
+      timeout: WAIT_TIMEOUT,
+    });
   });
 
   it('When action type is "webauthnCreate" and webauthnTransactionId is missing should log an error ', async () => {