+++ packages/libs/sdk-mixins/CHANGELOG.md
@@ -2,6 +2,28 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.2.0](https://github.com/descope/descope-js/compare/sdk-mixins-0.1.23...sdk-mixins-0.2.0) (2024-04-11)
+
+### Dependency Updates
+
+* `sdk-helpers` updated to version `0.1.20`
+* `sdk-component-drivers` updated to version `0.2.0`
+
+### Features
+
+* Audit widget RELEASE ([#502](https://github.com/descope/descope-js/issues/502)) ([d04199b](https://github.com/descope/descope-js/commit/d04199bc0d99083202f19cd28f0c2316cb19eb94))
+
+## [0.1.23](https://github.com/descope/descope-js/compare/sdk-mixins-0.1.22...sdk-mixins-0.1.23) (2024-04-09)
+
+### Dependency Updates
+
+* `sdk-helpers` updated to version `0.1.19`
+* `sdk-component-drivers` updated to version `0.1.19`
+
+### Bug Fixes
+
+* issue 6179 RELEASE ([#499](https://github.com/descope/descope-js/issues/499)) ([8d9fa34](https://github.com/descope/descope-js/commit/8d9fa34e20b9e018436be10eaa932621d7c7166e))
+
 ## [0.1.22](https://github.com/descope/descope-js/compare/sdk-mixins-0.1.21...sdk-mixins-0.1.22) (2024-04-08)
 
 ### Dependency Updates