+++ packages/web-component/CHANGELOG.md
@@ -2,6 +2,41 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.10.0](https://github.com/descope/descope-js/compare/web-component-3.9.0...web-component-3.10.0) (2024-04-15)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.10.34`
+
+### Features
+
+* add open in a new tab ([#519](https://github.com/descope/descope-js/issues/519)) RELEASE ([dec2bd7](https://github.com/descope/descope-js/commit/dec2bd7ad0167f1bc51f60db90e76d535b9d66e2))
+
+## [3.9.0](https://github.com/descope/descope-js/compare/web-component-3.8.40...web-component-3.9.0) (2024-04-12)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.10.33`
+
+### Features
+
+* NOTP ([#505](https://github.com/descope/descope-js/issues/505)) RELEASE ([c1f1d79](https://github.com/descope/descope-js/commit/c1f1d79b311532fec6d5779dbdadf3f239b7df46))
+
+## [3.8.40](https://github.com/descope/descope-js/compare/web-component-3.8.39...web-component-3.8.40) (2024-04-11)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.10.32`
+## [3.8.39](https://github.com/descope/descope-js/compare/web-component-3.8.38...web-component-3.8.39) (2024-04-10)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.10.31`
+## [3.8.38](https://github.com/descope/descope-js/compare/web-component-3.8.37...web-component-3.8.38) (2024-04-09)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.10.30`
 ## [3.8.37](https://github.com/descope/descope-js/compare/web-component-3.8.36...web-component-3.8.37) (2024-04-08)
 
 ### Dependency Updates