+++ packages/web-js-sdk/CHANGELOG.md
@@ -2,6 +2,33 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.10.34](https://github.com/descope/descope-js/compare/web-js-sdk-1.10.33...web-js-sdk-1.10.34) (2024-04-15)
+
+### Dependency Updates
+
+* `core-js-sdk` updated to version `2.14.0`
+## [1.10.33](https://github.com/descope/descope-js/compare/web-js-sdk-1.10.32...web-js-sdk-1.10.33) (2024-04-12)
+
+### Dependency Updates
+
+* `core-js-sdk` updated to version `2.13.0`
+## [1.10.32](https://github.com/descope/descope-js/compare/web-js-sdk-1.10.31...web-js-sdk-1.10.32) (2024-04-11)
+
+### Dependency Updates
+
+* `core-js-sdk` updated to version `2.12.2`
+## [1.10.31](https://github.com/descope/descope-js/compare/web-js-sdk-1.10.30...web-js-sdk-1.10.31) (2024-04-10)
+
+### Dependency Updates
+
+* `core-js-sdk` updated to version `2.12.1`
+
+### Bug Fixes
+
+* do not clear tokens for management api RELEASE ([#503](https://github.com/descope/descope-js/issues/503)) ([e615427](https://github.com/descope/descope-js/commit/e615427031e7ddd02d9507fb941f11ec83cb5616))
+
+## [1.10.30](https://github.com/descope/descope-js/compare/web-js-sdk-1.10.29...web-js-sdk-1.10.30) (2024-04-09)
+
 ## [1.10.29](https://github.com/descope/descope-js/compare/web-js-sdk-1.10.28...web-js-sdk-1.10.29) (2024-04-08)
 
 ## [1.10.28](https://github.com/descope/descope-js/compare/web-js-sdk-1.10.27...web-js-sdk-1.10.28) (2024-04-05)